requirements.txt
pyyaml==3.12
RPi.GPIO==0.6.3
requests==2.18.4
